Added static/ folder to .gitignore
Using bulma has changed the way we will use static files. Namely,
now each Django app has a static folder which is populated with
the resources needed for that app. Then, to gather all static
content to be served, run "python manage.py collectstatic". This
is both a side-effect of bulma and the appropriate way to serve
static content, so we're going to live with it :)