Bulma styling and UI improvements (#19)
* Removed unnecessary package-lock.json
* Whole lotta bulma restyling, keeping the feel of what Spenser did
* Added /static/ folder to .gitignore
* Using bulma has changed the way we will use static files. Namely,
now each Django app has a static folder which is populated with
the resources needed for that app. Then, to gather all static
content to be served, run "python manage.py collectstatic". This
is both a side-effect of bulma and the appropriate way to serve
static content, so we're going to live with it :)
* Readded catalog icon hovering
* Quality of life UI improvements
* Added precontent block for out-of-container elements like hero
banners and others. Made the file uploader automatic instead
of requiring to click submit.
* Fixed some misc errors/formatting issues
* UI catalog is now mobile friendly
* Removed debug logging from code
* Removed filename from upload button